echodune/requirements.txt
Django==4.2.16 
djangorestframework==3.15.2 
python-social-auth==0.3.6 
social-auth-app-django==5.4.2 
djangorestframework-simplejwt==5.3.1 
python-dotenv==1.0.1 
requests==2.32.3
Pillow==11.2.1
mutagen==1.47.0
django-cors-headers==4.7.0
social-auth-core==4.6.1
requests-oauthlib==2.0.0
oauthlib==3.3.1
